C++'ifying code initial checkpoint.
As it turns out, someone else went through the exercise of C++'ifying the tutorial during the same time as I was following along the BP tutorial: https://github.com/Olliebrown/LearningAgentsDriveCPP.

Which is proving to be a useful reference ("start with something that works"). However, for my exercise I'm going in a few different directions.